% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoaxis.R
\name{autoaxis}
\alias{autoaxis}
\title{Auto Axis Tool}
\usage{
autoaxis(
  side,
  major = NA,
  minor = NA,
  spacing = TRUE,
  format = "\%Y-\%m-\%d",
  tck = -0.03,
  major_grid = FALSE,
  minor_grid = FALSE,
  ...
)
}
\description{
Overlay base plot axis with custom intervals
}
\details{
Major and minor tick marks can be specified in a number of ways:
\itemize{
\item As a character string if the axis is datetime, such as 'year' or 'hour'
which are passed to seq().
\item As a tick interval using the default \code{spacing = TRUE}
\item As an approximate number of tick marks to include, using pretty() to find
the best interval, using \code{spacing = FALSE}
}

Major adds labels and ticks, minor is just half-sized ticks marks. Both
tick sizes can be changed (or direction changed) using \code{tck}.

Three different datetime axis are possible: year, day-offset, seconds-offset. Use
\code{format} to specify how the label should appear, such as '\%b \%Y' (see ?strptime)
\itemize{
\item Year should be treated as a conventional numeric axis, use major=1/12 not major='month'.
\item day-offset is an axis of class(x) = 'Date' and is identified if the axis range exists
within +/-9e4, or within dates 1723 - 2216.
\item second-offset is an axis of class(x)='POSIXct' and is identified by a range outside
of +/-9e4. This will give very strange results if your POSIXct axis is within
24 hours of 1970-01-01.
}

A grid can be added at the same time by setting major_grid or minor_grid to \code{TRUE}
or a colour. If \code{TRUE}, a transparent black is used by default.

Any other options can be passed through to axis() directly (see ?axis), most
notably \code{las = 2} to rotate the labels.
}
\examples{
plot(sunspots) # These time series are actually given in decimal years
  autoaxis(side=3, major=50, minor=10, spacing=T, major_grid='coral', minor_grid=T)
  autoaxis(side=2, minor=5) #
  autoaxis(side=4, major=11, minor=25, spacing=F, las=2, cex.axis=0.5, tck=0.02)

plot(seq(as.POSIXct('2020-01-01'),as.POSIXct('2020-01-03'),length.out=1e3), rnorm(1e3), xlab='POSIXct', xaxt='n')
  autoaxis(side=1, major='day', minor='hour', format = '\%A')

plot(seq(as.Date('2013-02-01'),as.Date('2020-01-03'),length.out=1e3), rnorm(1e3), xlab='Date', xaxt='n')
  autoaxis(side=1, major='year', minor='quarter', format = '\%Y')
}
