% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autopal.R
\name{autoimcol}
\alias{autoimcol}
\title{Create Capped Colour Palette For Images}
\usage{
autoimcol(z, set, limits = NA, n = 30, legend_len = 6, bias = 1)
}
\description{
Returns a colour palette and necessary breaks for image
}
\examples{
z = matrix(runif(100), nrow=10)
image(z, col = autoimcol(z = z, set = 'viridis', limits = c(0.1,0.9) ), breaks = .auto_imbreaks)
autolegend()
}
