% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autopal.R
\name{autolegend}
\alias{autolegend}
\title{Add Auto-Generated Legend}
\usage{
autolegend(...)
}
\arguments{
\item{...}{Arguments passed directly to \code{legend} -- legend text and fill are taken
automatically from hidden \code{.autocol_legend}}
}
\description{
Add a legend for the last autocol() set generated
}
\details{
If no location (such as 'bottom') is given, then it calls the locator() crosshairs
so the position of the legend can be picked interactively. All arguments are passed
to legend(), see ?legend for a full list.

Legend labels and fill are both generated by autopal() and autocol() and stored in
the hidden object \code{.autocol_legend} which can be manipulated if needs be.

See examples in ?autocol for plot + legend.
}
\examples{
# Simplest version: click-to-draw with locator()
plot(1:10, pch=16, col=autocol(1:10, 'Blues', legend_len = 5))
# autolegend() # Try me!

# Other neat versions -- note ?legend
autolegend(x=6, y=4, ncol=2, title='Draw at (6,4)')
autolegend('topleft', title='topleft', ncol=2, bty='n')
autolegend('bottom', inset=1, horiz=TRUE, bty='n')

# Manipulate the legend text, for example with format()
heatmap(as.matrix(eurodist), col=autopal('turbo', limits=range(eurodist)) )
.autocol_legend[[1]] = format(.autocol_legend[[1]],big.mark=',')
autolegend('bottom', inset=1, horiz=TRUE, title='Misleading miles between cities')

}
