% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autopal.R
\name{autolegend}
\alias{autolegend}
\title{Add Auto-Generated Legend}
\usage{
autolegend(...)
}
\arguments{
\item{...}{Arguments passed directly to \code{legend} -- legend text and colours are taken
automatically from \code{options('autolegend')}. See examples for useful
parameters, including \code{pch} and \code{pt.cex}}
}
\description{
Add a legend for the last autocol() set generated
}
\details{
If no location (such as 'bottom', or an x,y coordinate) is given, then it
calls the locator() crosshairs so the position of the legend can be picked
interactively. All arguments are passed to legend(), see ?legend for a full
list.

Legend labels and fill are both generated by autopal() and autocol() and stored
in the global "options('autolegend')" and can be manipulated if needs be.

See more examples in ?autocol for plot + legend.
}
\examples{
# Simplest version: click-to-draw with locator()
plot(1:10, pch=16, col=autocol(1:10, 'Blues', legend_len=5))
# autolegend() # Try me!

# Other neat versions -- note ?legend
autolegend(x=6, y=4, ncol=2, title='Draw at (6,4)')
autolegend('topleft', title='"topleft"', ncol=2, bty='n')
autolegend('bottom', inset=1, horiz=TRUE, bty='n', title='Above plot')

# Use pch (and optionally pt.cex) in legend -- these get recycled
autolegend('bottom', horiz=TRUE, pch=16, pt.cex=3, title='pch=16, pt.cex=3')
autolegend('right', pch=1:10, pt.cex=2, title='pch=1:10')

# Manipulate the legend text, for example with format(), this is a bit long-winded!
heatmap(as.matrix(eurodist), col=autopal('turbo', limits=range(eurodist)) )
current_legend = options('autolegend')[[1]]
options(autolegend = list(format(current_legend[[1]], big.mark=','), current_legend[[2]]))
autolegend('bottom', inset=1, horiz=TRUE, title='Misleading miles between cities')

# No helper exists yet for creating size or shape legends -- follow this idea...
with(airquality, plot(Temp, pch=16, cex=Solar.R/100, col=autocol(Ozone, set='Reds')))
cex_legend = pretty(airquality$Solar.R)
legend('bottom', pt.cex=cex_legend/100, legend=cex_legend, pch=1,
  horiz=TRUE, title='Solar.R', bty='n' )
autolegend('bottom', inset=1, horiz=TRUE, title='Ozone', bty='n')

}
