% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autopal.R
\name{autolegend}
\alias{autolegend}
\title{Add Auto-Generated Legend}
\usage{
autolegend(...)
}
\arguments{
\item{...}{Arguments passed directly to \code{legend} - legend text and fill are taken automatically from hidden \code{.autocol_legend}}
}
\description{
Add a legend for the last autocol() set generated
}
\details{
If no location (such as 'bottom') is given, then it calls the locator() crosshairs
so the position of the legend can be picked interactively.
}
\examples{
autolegend('topright', ncol = 2, title = 'Legend')
autolegend(horiz = T, bty = 'n') # Try clicking just under the plot title

}
