% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autozoom.R
\name{autozoom}
\alias{autozoom}
\title{Auto Zoom to Create a Dynamic Plot}
\usage{
autozoom(plt)
}
\description{
Takes a plot function as its argument and repeats the call every time the plot
is interacted with (click to zoom), using the new plot limits.
}
